Refactor : #16 GuestHome, GuestWrite url 변경
