--- src/pages/GuestLayout.tsx
@@ -1,265 +1,0 @@
-import React, { useEffect, useState, useRef, useCallback } from 'react';
-import styled from 'styled-components';
-import Draggable from 'react-draggable';
-import { useNavigate } from 'react-router-dom';
-import axios from 'axios';
-
-import theme from '../common/style/theme';
-import { Title, Button } from '../components/common/style';
-import BackBtn from '../components/common/BackBtn';
-import { flowerwraps, leaves, letters } from '../constants';
-
-const MainWrapper = styled.div`
-  width: 100%;
-  min-height: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  padding: 45px 15px 0 15px;
-`;
-
-// !isLayout
-const ContentWrapper = styled.form`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 15px;
-`;
-
-const NicknameInput = styled.input`
-  width: 100%;
-  border-radius: 12px;
-  height: 40px;
-  flex-shrink: 0;
-  font-size: ${theme.calcRem(24)};
-  padding-left: 6px;
-  &:focus {
-    outline: none;
-  }
-`;
-
-const LetterWrapper = styled.div`
-  height: 100%;
-  width: 100%;
-  position: relative;
-  margin-top: 15px;
-`;
-
-const Letter = styled.textarea`
-  width: 100%;
-  height: 100%;
-  font-size: ${theme.calcRem(24)};
-  padding: 30px;
-  border: none;
-  z-index: 1;
-  position: absolute;
-  top: 0;
-  left: 0;
-  background-color: transparent;
-  &:focus {
-    border: none;
-    outline: none;
-  }
-`;
-
-const LetterImg = styled.img`
-  width: 100%;
-  height: 100%;
-  opacity: 0.4;
-  position: absolute;
-  top: 0;
-  left: 0;
-`;
-
-// isLayout
-const MainTitle = styled(Title)`
-  flex-shrink: 0;
-  margin-top: 15px;
-`;
-
-const ImgWrapper = styled.div`
-  height: 100%;
-  width: 100%;
-  position: relative;
-`;
-
-const UserImg = styled.img`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-`;
-
-const SelectedLeaf = styled.img`
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 999;
-  width: 100px;
-  height: 100px;
-  cursor: move;
-`;
-
-const PrevBtn = styled.a`
-  width: 20px;
-  height: 20px;
-  border-top: 3.5px solid #000;
-  border-left: 3.5px solid #000;
-  transform: rotate(-45deg);
-  flex-shrink: 0;
-`;
-
-const NextBtn = styled(Button)`
-  margin-top: 15px;
-  flex-shrink: 0;
-`;
-
-interface IPos {
-  x: number;
-  y: number;
-}
-
-interface IBox {
-  height: number;
-  width: number;
-}
-
-export default function GuestLayout() {
-  const [isLayout, setIsLayout] = useState(false);
-  const [nickname, setNickname] = useState('');
-  const [text, setText] = useState('');
-  const [pos, setPos] = useState<IPos>({ x: 0, y: 0 });
-  const [originPos, setOriginPos] = useState<IPos>({ x: 0, y: 0 });
-  const [box, setBox] = useState<IBox>({ width: 0, height: 0 });
-
-  const navigate = useNavigate();
-  const imgWrapper = useRef() as React.MutableRefObject<HTMLDivElement>;
-  const nodeRef = useRef(null);
-  const categoryId = 1;
-  const imgSize = 100;
-
-  useEffect(() => {
-    if (isLayout) {
-      const wrapperBox = imgWrapper.current.getBoundingClientRect();
-      setBox({ ...box, width: wrapperBox.width, height: wrapperBox.height });
-      console.log(wrapperBox);
-    }
-
-    // const fetchData = async () => {
-    //   const response = await axios.get(`${process.env.apiUrl}/api/post/${1}`);
-    // };
-    // fetchData();
-  }, [isLayout]);
-
-  const startPos = useCallback(
-    (data: any) => {
-      if (
-        data.x >= 0 &&
-        data.x + imgSize * 0.66 < box.width &&
-        data.y >= 0 &&
-        data.y + imgSize * 0.66 < box.height
-      ) {
-        setOriginPos({ ...originPos, x: data.x, y: data.y });
-      } else {
-        setPos({ ...pos, x: originPos.x, y: originPos.y });
-      }
-    },
-    [pos, originPos, box]
-  );
-
-  const trackPos = useCallback(
-    (data: any) => {
-      setPos({ ...pos, x: data.x, y: data.y });
-    },
-    [pos]
-  );
-
-  const stopPos = useCallback(
-    (data: any) => {
-      if (
-        data.x >= 0 &&
-        data.x + imgSize * 0.66 < box.width &&
-        data.y >= 0 &&
-        data.y + imgSize * 0.66 < box.height
-      ) {
-        setPos({ ...pos, x: data.x, y: data.y });
-      } else {
-        setPos({ ...pos, x: originPos.x, y: originPos.y });
-      }
-    },
-    [pos, originPos, box]
-  );
-
-  const onNicknameChange = useCallback(
-    (e: React.FormEvent<HTMLInputElement>) => {
-      const { value } = e.currentTarget;
-      setNickname(value);
-    },
-    [nickname]
-  );
-
-  const onTextChange = useCallback(
-    (e: React.FormEvent<HTMLTextAreaElement>) => {
-      const { value } = e.currentTarget;
-      setText(value);
-    },
-    [text]
-  );
-
-  const onPrevClick = useCallback(() => {
-    setIsLayout(false);
-  }, []);
-
-  const onNextClick = useCallback((e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setIsLayout(true);
-  }, []);
-
-  const onFinishClick = useCallback(() => {
-    const xRatio = pos.x / box.width;
-    const yRatio = pos.y / box.height;
-    console.log(pos, box, xRatio, yRatio);
-    // navigate('/guest/layout');
-  }, [pos, box]);
-
-  return (
-    <MainWrapper>
-      {isLayout ? <PrevBtn onClick={onPrevClick} /> : <BackBtn />}
-      {isLayout && <MainTitle>원하는 위치에 꽃을 배치해주세요.</MainTitle>}
-      {!isLayout ? (
-        <ContentWrapper onSubmit={onNextClick}>
-          <NicknameInput
-            placeholder="닉네임 (최대 8자)"
-            required
-            maxLength={8}
-            onChange={onNicknameChange}
-            value={nickname}
-          />
-          <LetterWrapper>
-            <Letter required onChange={onTextChange} value={text} />
-            <LetterImg src={letters[categoryId]} />
-          </LetterWrapper>
-          <NextBtn>다음</NextBtn>
-        </ContentWrapper>
-      ) : (
-        <ImgWrapper ref={imgWrapper}>
-          <UserImg src={flowerwraps[categoryId]} />
-          <Draggable
-            position={pos}
-            onStart={(e, data) => startPos(data)}
-            onDrag={(e, data) => trackPos(data)}
-            onStop={(e, data) => stopPos(data)}
-            nodeRef={nodeRef}
-          >
-            <SelectedLeaf src={leaves[categoryId][0].url} ref={nodeRef} />
-          </Draggable>
-        </ImgWrapper>
-      )}
-
-      {isLayout && <NextBtn onClick={onFinishClick}>완료</NextBtn>}
-    </MainWrapper>
-  );
-}
